+++ pomdp_py/problems/tiger/__init__.py
@@ -1,1 +1,1 @@
-from .tiger_problem import TigerProblem, test_planner
+from .tiger_problem import TigerProblem, test_planner, make_tiger
